added taz customization helper, refactored/moved code
