+++ taz.neo/ViewController/SettingsVC.swift
@@ -977,124 +977,4 @@ class SimpleHeaderView: UIView,  UIStyleChangeDelegate{
   }
 }
 
-// MARK: - ext. UIView (Wrapper)
-extension UIView {
-  
-  /// add self to a wrapper View, pin with dist and return wrapper view
-  /// - Parameter dist: dist to pin between wrapper and self
-  /// - Returns: wrapper
-  @discardableResult
-  func wrapper(_ insets: UIEdgeInsets = Const.Insets.Small) -> UIView {
-    let wrapper = UIView()
-    wrapper.addSubview(self)
-    pin(self.left, to: wrapper.left, dist: insets.left)
-    pin(self.right, to: wrapper.right, dist: insets.right)
-    pin(self.top, to: wrapper.top, dist: insets.top)
-    pin(self.bottom, to: wrapper.bottom, dist: insets.bottom)
-    return wrapper
-  }
-  
-  /// set backgroundColor and return self (for chaining)
-  /// - Parameter backgroundColor: backgroundColor to set
-  /// - Returns: self
-  @discardableResult
-  func set(backgroundColor: UIColor) -> UIView {
-    self.backgroundColor = backgroundColor
-    return self
-  }
-  
-//  ///Blur Idea from: https://stackoverflow.com/questions/30953201/adding-blur-effect-to-background-in-swift
-//  /// not working here for chaining, need also effect style depending dark/light
-//  func addBlur() -> UIView {
-//    let blurEffect = UIBlurEffect(style: UIBlurEffect.Style.extraLight)
-//    let blurEffectView = UIVisualEffectView(effect: blurEffect)
-//    blurEffectView.frame = self.bounds
-//    blurEffectView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-//    self.addSubview(blurEffectView)
-//    return self
-//  }
-}
-
-
-extension ButtonControl {
-  
-  @discardableResult
-  func tazX(_ isUpdate:Bool = false) -> Self {
-    guard let bv = self as? Button<CircledXView> else { return self }
-    bv.buttonView.circleColor = Const.SetColor.ios(.secondarySystemFill).color
-    bv.buttonView.color = Const.SetColor.ios(.secondaryLabel).color
-    bv.buttonView.activeColor = Const.SetColor.ios(.secondaryLabel).color.withAlphaComponent(0.1)
-    if isUpdate ==  true { return self }
-    self.pinHeight(35)
-    self.pinWidth(35)
-    bv.buttonView.isCircle = true
-    bv.buttonView.innerCircleFactor = 0.5
-    return self
-  }
-  
-  @discardableResult
-  func tazButton(_ isUpdate:Bool = false) -> Self {
-    guard let bv = self as? Button<TextView> else { return self }
-    bv.buttonView.circleColor = Const.SetColor.ios(.secondarySystemFill).color
-    bv.buttonView.label.textColor = Const.SetColor.ios(.secondaryLabel).color
-    bv.buttonView.color = Const.SetColor.ios(.secondaryLabel).color
-    bv.buttonView.activeColor = Const.SetColor.ios(.secondaryLabel).color.withAlphaComponent(0.1)
-    if isUpdate ==  true { return self }
-    self.pinSize(CGSize(width: 28, height: 28), priority: .defaultHigh)
-    bv.buttonView.isCircle = true
-    bv.buttonView.font
-    = Const.Fonts.contentFont(size: Const.Size.DefaultFontSize)//16
-    return self
-  }
-}
-
-extension UIView {
-  
-  func inset(insets: UIEdgeInsets) {
-    guard let superview = self.superview else { return }
-    
-    self.leftAnchor.constraint(equalTo: superview.leftAnchor,
-                               constant: insets.left,
-                               priority: .defaultHigh,
-                               activate: true)
-    self.rightAnchor.constraint(equalTo: superview.rightAnchor, constant: -insets.right,
-                                priority: .defaultHigh,
-                                activate: true)
-    self.topAnchor.constraint(equalTo: superview.topAnchor, constant: insets.top,
-                              priority: .defaultHigh,
-                              activate: true)
-    self.bottomAnchor.constraint(equalTo: superview.bottomAnchor, constant: -insets.bottom,
-                                 priority: .defaultHigh,
-                                 activate: true)
-  }
-}
-
-extension NSLayoutAnchor {
-  @discardableResult
-  @objc open func constraint(equalTo anchor: NSLayoutAnchor<AnchorType>, constant c: CGFloat, priority: UILayoutPriority, activate: Bool = false) -> NSLayoutConstraint {
-    let constraint = self.constraint(equalTo: anchor, constant: c)
-    constraint.priority = priority
-    if activate { constraint.isActive = true}
-    return constraint
-  }
-}
-
 class FooterView: UITableViewHeaderFooterView{}
-
-extension Array {
-  /// The  second element of the collection.
-  ///
-  /// If count is less then 2, the value of this property is `nil`.
-  ///
-  ///     let numbers = [10, 20, 30, 40, 50]
-  ///     if let firstNumber = numbers.first {
-  ///         print(firstNumber)
-  ///     }
-  ///     // Prints "20"
-  @inlinable public var  second: Element? {
-    get {
-      if self.count > 1 { return self[1] }
-      return nil
-    }
-  }
-}
